requirements.txt
django==3.1
djangorestframework==3.12.4
Pillow==8.3.2